requirements.txt
pandas
numpy
scikit-learn
joblib
pyyaml
pytest